requirements.txt
pyserial~=3.5
mss~=9.0.1
numpy~=1.25.0
opencv-python~=4.7.0.72